fix: modified core entities

+++ core/src/main/java/ru/amorozov/domain/entities/ContentType.java
@@ -7,7 +7,7 @@ import ru.amorozov.domain.entities.enums.Type;
  * @author Aleksey Morozov
  * @since 25.08.2022
  */
-public record ContentType(int id,
+public record ContentType(Long id,
                           @NonNull Type name) implements BasicEntity {
 
 }
